Implement concurrent transactions using iris.
TDB does not support more than a single transaction at a time. Therefore,
we use iris to queue operations for a given transaction until the commit
for that transaction comes in. Upon that time, all of the operations for
the transaction can happen serially while within the exclusive handler.

These will not currently work with synchronous transactions, which is why
they are not included in this patch.  To support synchronous transactions
we need to begin applying operations for a transaction before the commit
has been called.  This will require queing non-transaction messages until
the transaction has completed.